Adaptions for IGC setup of Flarm
git-svn-id: https://svn.kflog.org/svn/repos/cumulus/qt4/trunk@5360 c8f415cb-fa42-0410-b21e-89df95f64ae5

+++ gpsClient/gpsclient.cpp
@@ -268,7 +268,7 @@ int GpsClient::writeGpsData( const char *sentence )
       return -1;
     }
 
-  uint csum = calcCheckSum( sentence );
+  uchar csum = calcCheckSum( sentence );
   QString check;
   check.sprintf ("*%02X\r\n", csum);
   QString cmd (sentence + check);
